+++ dev/src/app/(payload)/admin/importMap.js
@@ -23,9 +23,10 @@ import { ReindexButton as ReindexButton_aead06e4cbf6b2620c5c51c9ab283634 } from 
 import { FolderTypeField as FolderTypeField_3817bf644402e67bfe6577f60ef982de } from '@payloadcms/ui'
 import { RowLabel as RowLabel_ec255a65fa6fa8d1faeb09cf35284224 } from '@/Header/RowLabel'
 import { RowLabel as RowLabel_1f6ff6ff633e3695d348f4f3c58f1466 } from '@/Footer/RowLabel'
-import { default as default_00089f22fccacdde213d7918ba8947fd } from '@/components/SSOLoginButton'
+import { default as default_624daa126368b3ecb5ff0779f4d41267 } from 'payload-auth-cookie/admin/DefaultSSOLoginButton'
 import { default as default_1a7510af427896d367a49dbf838d2de6 } from '@/components/BeforeDashboard'
 import { default as default_8a7ab0eb7ab5c511aba12e68480bfe5e } from '@/components/BeforeLogin'
+import { default as default_169241da18aab009a8d9815ba22c176e } from 'payload-auth-cookie/admin/DefaultSSOLogoutRedirect'
 
 export const importMap = {
   "@payloadcms/richtext-lexical/rsc#RscEntryLexicalCell": RscEntryLexicalCell_44fe37237e0ebf4470c9990d8cb7b07e,
@@ -53,7 +54,8 @@ export const importMap = {
   "@payloadcms/ui#FolderTypeField": FolderTypeField_3817bf644402e67bfe6577f60ef982de,
   "@/Header/RowLabel#RowLabel": RowLabel_ec255a65fa6fa8d1faeb09cf35284224,
   "@/Footer/RowLabel#RowLabel": RowLabel_1f6ff6ff633e3695d348f4f3c58f1466,
-  "@/components/SSOLoginButton#default": default_00089f22fccacdde213d7918ba8947fd,
+  "payload-auth-cookie/admin/DefaultSSOLoginButton#default": default_624daa126368b3ecb5ff0779f4d41267,
   "@/components/BeforeDashboard#default": default_1a7510af427896d367a49dbf838d2de6,
-  "@/components/BeforeLogin#default": default_8a7ab0eb7ab5c511aba12e68480bfe5e
+  "@/components/BeforeLogin#default": default_8a7ab0eb7ab5c511aba12e68480bfe5e,
+  "payload-auth-cookie/admin/DefaultSSOLogoutRedirect#default": default_169241da18aab009a8d9815ba22c176e
 }